# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: OMNIRA API
  version: 0.2.0
  description: |
    OMNIRA SaaS Platform API — R0.2 Extended Features

    Multi-tenant platform with Row-Level Security, shared strong isolation,
    event-driven architecture, and cursor-based pagination.
  contact:
    name: OMNIRA
  license:
    name: Proprietary

servers:
  - url: http://localhost:8080
    description: Development
  - url: https://api.omnira.local
    description: Production

tags:
  - name: Health & Observability
    description: System health checks and metrics
  - name: Tenancy
    description: Tenant management and membership
  - name: Internal
    description: Internal endpoints (not exposed publicly)

paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Health check endpoint
      tags:
        - Health & Observability
      description: |
        Comprehensive health check including database and NATS connectivity.
        Returns 200 if healthy, 503 if any component is unhealthy.
      responses:
        '200':
          description: All systems healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '503':
          description: One or more components unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitExceeded'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'

  /metrics:
    get:
      operationId: getMetrics
      summary: Prometheus metrics
      tags:
        - Health & Observability
      description: |
        System metrics in Prometheus text format (version 0.0.4).
        Includes health status and latency per component.
      responses:
        '200':
          description: Prometheus metrics
          content:
            text/plain:
              schema:
                type: string
              example: |
                # HELP omnira_health_status Health status of components
                # TYPE omnira_health_status gauge
                omnira_health_status{component="database"} 1
                omnira_health_status{component="nats"} 1
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitExceeded'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'

  /tenant:
    get:
      operationId: getTenantMe
      summary: Get current tenant
      tags:
        - Tenancy
      description: |
        Retrieves the current tenant associated with the authenticated user.
        Requires valid JWT Bearer token.
      security:
        - Bearer: []
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '401':
          description: Unauthorized (missing or invalid token)
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '403':
          description: Forbidden (user not authorized for tenant)
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitExceeded'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'

  /tenant/memberships:
    get:
      operationId: listMemberships
      summary: List tenant memberships
      tags:
        - Tenancy
      description: |
        Lists all memberships in the current tenant.
        Requires authentication and active membership.
      security:
        - Bearer: []
      responses:
        '200':
          description: Memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipsResponse'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '401':
          description: Unauthorized
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '403':
          description: Forbidden
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitExceeded'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'

  /tenant/memberships/{userID}:
    post:
      operationId: createMembership
      summary: Grant membership
      tags:
        - Tenancy
      description: |
        Grants a user membership in the current tenant with specified role.
        Requires admin-level access.
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipRequest'
      responses:
        '201':
          description: Membership granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '400':
          description: Invalid request
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '401':
          description: Unauthorized
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '403':
          description: Forbidden
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitExceeded'
          headers:
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'

    delete:
      operationId: revokeMembership
      summary: Revoke membership
      tags:
        - Tenancy
      description: |
        Revokes a user's membership in the current tenant.
        Requires admin-level access.
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Membership revoked
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Membership not found

  /internal/health/live:
    get:
      operationId: getLivenessProbe
      summary: Kubernetes liveness probe
      tags:
        - Internal
      description: Simple liveness probe for Kubernetes health checks.
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /internal/health/ready:
    get:
      operationId: getReadinessProbe
      summary: Kubernetes readiness probe
      tags:
        - Internal
      description: |
        Readiness probe that checks if service is ready to accept traffic.
        Checks dependent services (database, cache, etc).
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready
        '503':
          description: Service not ready

components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RSA-signed JWT token in Authorization header.
        Format: Authorization: Bearer {token}

  headers:
    X-RateLimit-Remaining:
      description: Number of requests remaining in current rate limit window
      schema:
        type: integer
        example: 42

    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        format: int64
        example: 1695312345

    Retry-After:
      description: Seconds to wait before retrying (when rate limited)
      schema:
        type: integer
        example: 45

  parameters:
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (default 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

    CursorParam:
      name: cursor
      in: query
      description: Cursor for pagination (empty for first page)
      schema:
        type: string
        example: "dXNlci0xOjE2OTUzMTIzNDU="

    SortParam:
      name: sort
      in: query
      description: Sort specification (format field:asc or field:desc)
      schema:
        type: string
        example: "created_at:desc"

  schemas:
    RateLimitExceeded:
      type: object
      required:
        - error
        - retry_after
      properties:
        error:
          type: string
          example: Rate limit exceeded
        retry_after:
          type: integer
          description: Seconds to wait before retrying
          example: 45
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - components
      properties:
        status:
          type: string
          enum:
            - healthy
            - unhealthy
            - unknown
        timestamp:
          type: string
          format: date-time
        components:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ComponentHealth'

    ComponentHealth:
      type: object
      required:
        - name
        - status
      properties:
        name:
          type: string
          example: database
        status:
          type: string
          enum:
            - healthy
            - unhealthy
            - unknown
        message:
          type: string
          example: ping failed
        latency_ms:
          type: integer
          example: 5

    TenantResponse:
      type: object
      required:
        - id
        - legal_name
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        legal_name:
          type: string
          example: Acme Corp
        isolation_profile:
          type: string
          enum:
            - standard
            - premium
            - dedicated
        status:
          type: string
          enum:
            - active
            - inactive
            - suspended
        created_at:
          type: string
          format: date-time

    MembershipsResponse:
      type: object
      required:
        - memberships
      properties:
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/MembershipResponse'

    MembershipResponse:
      type: object
      required:
        - id
        - tenant_id
        - user_id
        - role_id
        - status
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role_id:
          type: string
          example: admin
        status:
          type: string
          enum:
            - active
            - inactive
            - revoked
        created_at:
          type: string
          format: date-time

    CreateMembershipRequest:
      type: object
      required:
        - role_id
      properties:
        role_id:
          type: string
          example: viewer
          description: System or tenant role identifier

    PageMetadata:
      type: object
      required:
        - count
        - limit
        - has_more
      properties:
        count:
          type: integer
          description: Number of items in this page
          example: 20
        limit:
          type: integer
          description: Maximum items per page
          example: 20
        has_more:
          type: boolean
          description: Whether more items exist
        next_cursor:
          type: string
          description: Cursor for next page (null if has_more is false)
          example: "dXNlci0xOjE2OTUzMTIzNDU="